src/main/webapp/app/entities/serviciosmedicos/motivo-referencia/route/motivo-referencia-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IMotivoReferencia } from '../motivo-referencia.model';
import { MotivoReferenciaService } from '../service/motivo-referencia.service';

import motivoReferenciaResolve from './motivo-referencia-routing-resolve.service';

describe('MotivoReferencia routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let service: MotivoReferenciaService;
  let resultMotivoReferencia: IMotivoReferencia | null | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    service = TestBed.inject(MotivoReferenciaService);
    resultMotivoReferencia = undefined;
  });

  describe('resolve', () => {
    it('should return IMotivoReferencia returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        motivoReferenciaResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMotivoReferencia = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMotivoReferencia).toEqual({ id: 123 });
    });

    it('should return null if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      TestBed.runInInjectionContext(() => {
        motivoReferenciaResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMotivoReferencia = result;
          },
        });
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultMotivoReferencia).toEqual(null);
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse<IMotivoReferencia>({ body: null })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      TestBed.runInInjectionContext(() => {
        motivoReferenciaResolve(mockActivatedRouteSnapshot).subscribe({
          next(result) {
            resultMotivoReferencia = result;
          },
        });
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultMotivoReferencia).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
